Общие функции перенесены в отдельные файлы
