+++ HW_2/project/include/multi/multi_process_utils.h
@@ -6,13 +6,10 @@
 #define HW_2_PROJECT_INCLUDE_MULTI_MULTI_PROCESS_UTILS_H_
 
 #include <stdio.h>
-#include "types.h"
+#include "utils.h"
 
 # define PARENT_PID -5
 
-void print_matrix(Matrix* matrix);
-Matrix* input_matrix();
-Matrix* read_file(const char* file_name);
 int create_forks(int num, int *pids);
 Calculation_res *create_shared_memory();
 Calculation_res* multi_process(char* file_name, int num_forks);
