+++ HW_2/project/include/one/one_proc_utils.h
@@ -1,0 +1,13 @@
+//
+// Created by Andrew Kireev on 18.10.2020.
+//
+
+#ifndef HW_2_PROJECT_INCLUDE_ONE_PROC_UTILS_H_
+#define HW_2_PROJECT_INCLUDE_ONE_PROC_UTILS_H_
+
+#include <stdio.h>
+#include "utils.h"
+
+Calculation_res calculate_matrix(Matrix* matrix);
+
+#endif  // HW_2_PROJECT_INCLUDE_ONE_PROC_UTILS_H_
